SWC-5019

+++ src/main/java/org/sagebionetworks/web/client/widget/evaluation/EvaluationEditorModalView.java
@@ -35,7 +35,7 @@ public interface EvaluationEditorModalView extends IsWidget {
 
 	void setRoundDuration(Long roundDurationMs);
 
-	Double getRoundDuration();
+	Long getRoundDuration();
 
 	void setNumberOfRounds(Long numberOfRounds);
